+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0]
+### Added
+- `TskVsPartHandle` to handle io for the partition volume
+
+### Fixed
+- Handling of null pointers returned from `tsk_error_get`
+
+### Changed
+- tests work of of test images instead of logical volume
+- renamed `TskImg::from_source` to `TskImg::from_utf8_sing` to better fit with libtsk function name
+
 ## [0.2.0]
 ### Added
 - `TskImg::from_tsk_img_info_ptr()` to create a TskImg from `NonNull<tsk::TSK_IMG_INFO>`